Move suggestions into a rattler package and treat the root as a group command
Addresses review feedback: the suggestion machinery moves to
pkg/cmd/pulumi/rattler next to constrictor, bare `pulumi` now prints help
but exits non-zero like any other group command, and the synonym table's
symmetry is documented and tested.
